+++ Boleto2.Net/Util/TTiposDadoEDI.cs
@@ -1,0 +1,77 @@
+namespace Boleto2Net
+{
+    /// <summary>
+    /// Representa cada tipo de dado possível em um arquivo EDI.
+    /// </summary>
+    public enum TTiposDadoEDI
+    { 
+        /// <summary>
+        /// Representa um campo alfanumérico, alinhado à esquerda e com brancos à direita. A propriedade ValorNatural é do tipo String
+        /// </summary>
+        ediAlphaAliEsquerda_____,
+        /// <summary>
+        /// Representa um campo alfanumérico, alinhado à direita e com brancos à esquerda. A propriedade ValorNatural é do tipo String
+        /// </summary>
+        ediAlphaAliDireita______,
+        /// <summary>
+        /// Representa um campo numérico inteiro alinhado à direita com zeros à esquerda. A propriedade ValorNatural é do tipo Int ou derivados
+        /// </summary>
+        ediInteiro______________,
+        /// <summary>
+        /// Representa um campo numérico com decimais, sem o separador de decimal. A propriedade ValorNatural é do tipo Double
+        /// </summary>
+        ediNumericoSemSeparador_,
+        /// <summary>
+        /// Representa um campo numérico com decimais, com o caracter ponto (.) como separador decimal,
+        /// alinhado à direita com zeros à esquerda. A propriedade ValorNatural é do tipo Double
+        /// </summary>
+        ediNumericoComPonto_____,
+        /// <summary>
+        /// Representa um campo numérico com decimais, com o caracter vírgula (,) como separador decimal,
+        /// alinhado à direita com zeros à esquerda. A propriedade ValorNatural é do tipo Double
+        /// </summary>
+        ediNumericoComVirgula___,
+        /// <summary>
+        /// Representa um campo de data no formato ddm/mm/aaaa. A propriedade ValorNatural é do tipo DateTime
+        /// </summary>
+        ediDataDDMMAAAA_________,
+        /// <summary>
+        /// Representa um campo de data no formato aaaa/mm/dd. A propriedade ValorNatural é do tipo DateTime
+        /// </summary>
+        ediDataAAAAMMDD_________,
+        /// <summary>
+        /// Representa um campo de data no formato dd/mm. A propriedade ValorNatural é do tipo DateTime, com o ano igual a 1900
+        /// </summary>
+        ediDataDDMM_____________,
+        /// <summary>
+        /// Representa um campo de data no formato mm/aaaa. A propriedade ValorNatural é do tipo DateTime, com o dia igual a 01
+        /// </summary>
+        ediDataMMAAAA___________,
+        /// <summary>
+        /// Representa um campo de data no formato mm/dd. A propriedade ValorNatural é do tipo DateTime com o ano igual a 1900
+        /// </summary>
+        ediDataMMDD_____________,
+        /// <summary>
+        /// Representa um campo de hora no formato HH:MM. A propriedade ValorNatural é do tipo DateTime, com a data igual a 01/01/1900
+        /// </summary>
+        ediHoraHHMM_____________,
+        /// <summary>
+        /// Representa um campo de hora no formato HH:MM:SS. A propriedade ValorNatural é do tipo DateTime, com a data igual a 01/01/1900
+        /// </summary>
+        ediHoraHHMMSS___________,
+        /// <summary>
+        /// Representa um campo de data no formato DD/MM/AAAA. A propriedade ValorNatural é do tipo DateTime.
+        /// </summary>
+        ediDataDDMMAA___________,
+        /// <summary>
+        /// Representa um campo de data no formato DD/MM/AAAA, porém colocando zeros no lugar de espaços no ValorFormatado. A propriedade
+        /// ValorNatural é do tipo DateTime, e este deve ser nulo caso queira que a data seja zero.
+        /// </summary>
+        ediDataDDMMAAAAWithZeros,
+        /// <summary>
+        /// Representa um campo de data no formato AAAA/MM/DD, porém colocando zeros no lugar de espaços no ValorFormatado. A propriedade
+        /// ValorNatural é do tipo DateTime, e este deve ser nulo caso queira que a data seja zero.
+        /// </summary>
+        ediDataAAAAMMDDWithZeros
+    }
+}
